Removing Grunt
Ok so being as this is a rails project I didn't see much need for grunt,
although as I wanted to experiment with it I did. Due to how this project
is setup related to assets I saw no need for anything else other than 'check
gems' overall it seems unnecessary to keep grunt in the repo or on this project.